requirements.txt
streamlit>=1.32.0
pypdf>=4.0.0
Pillow>=10.0.0
reportlab>=4.0.0
